+++ Documents/ChangeLog.md
@@ -1,5 +1,17 @@
 # Change Log
 
-- 2019.12.25 : Try Support Mip Levels. Update Document.
-- 2019.12.26 : Add `GpuTextureRef` interface. Update Document.
-- 2019.12.28 : Add `GpuTextureBuffer` interface. Update Document.
+## 2019.12.25
+
+- 2019.12.25 : Try Support Mip Levels. 
+- Update Document.
+
+## 2019.12.26
+
+- 2019.12.26 : Add `GpuTextureRef` interface. 
+- Update Document.
+
+## 2019.12.28
+
+- Add `GpuTextureBuffer` interface. 
+- Remove `GpuGraphicsCommandList::copyMemoryToXXX`.
+- Update Document.